Update README.md
Studies assessing the climate are essential for comprehending and combating climate change. In order to identify temperature trends, extreme weather patterns, and environmental effects, they analyse historical climate data. These initiatives influence public policy, stimulate innovation, and increase awareness. They are essential for coping with the repercussions of a changing climate and minimizing their effects, ensuring a sustainable future. A comprehensive tool for analysing the world's climate, the "GlobalLandTemperaturesByCountry" collection contains historical temperature records from numerous nations. This dataset covers several centuries and offers monthly temperature readings. It is derived from the Berkeley Earth Surface Temperature (BEST) project. Each item contains information on the nation, the date (in the YYYY-MM-DD format), the average land temperature for that month in degrees Celsius, and the measurement's associated uncertainty or margin of error. Isolation Forest was used to identify anomalies in the datasets of various nations, and the LSTM Model and ARIMA Model were used to foretell the pattern of the climate analysis.